+++ opi_midi_monitor/firmware/main.cpp
@@ -2,7 +2,7 @@
  * @file main.cpp
  *
  */
-/* Copyright (C) 2016-2025 by Arjan van Vught mailto:info@gd32-dmx.org
+/* Copyright (C) 2016-2026 by Arjan van Vught mailto:info@gd32-dmx.org
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -65,7 +65,7 @@ int main() { // NOLINT
 
     console::Clear();
     fw.Print();
-    printf("MIDI Monitor, baudrate : %d, interface : %s", midi.GetBaudrate(), EXT_MIDI_UART_NAME);
+    printf("MIDI Monitor, baudrate : %u, interface : %s", static_cast<unsigned>(midi.GetBaudrate()), EXT_MIDI_UART_NAME);
 
     MidiMonitor monitor;
     monitor.Init();
